pyproject.toml
[tool.poetry]
name = "oidc client"
version = "0.1.0"
description = "Magically transforms Papermerge into OIDC client"
authors = ["Eugen Ciur <eugen@papermerge.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.2"
python-jose = "^3.3.0"
pydantic-settings = "^2.2.1"
uvicorn = "^0.29.0"

[tool.poetry.group.dev.dependencies]
taskipy = "^1.12.2"

[tool.taskipy.tasks]
server = "uvicorn oidc_client.main:app --host 0.0.0.0 --port 8010"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
